--- script/bataille_navale.py
@@ -1,50 +1,0 @@
-# Importation des fonctions depuis action.py
-from action import *
-
-# Importation permettant de récuperer les arguments utilisé lors de l'appel du code
-import sys
-
-
-# Initialisation des valeur par défaut
-partie_finie = False
-partie_gagnee = False
-tour_joueur = 0
-type_de_partie = "console"
-
-# Fonction qui se lance lors du lancement du script
-if __name__ == '__main__':
-    # Vérification du nombre d'arguments nécessaire
-    if len(sys.argv) < 2:
-        print("Précisez une un mode de jeu en paramètre\nconsole/gui")
-        sys.exit(1)
-
-    argument = sys.argv[1]
-
-    if argument == "console":
-        print("On démarre en mode console")
-        # Attribution du style de bataille (nombre de colonne et de ligne)
-        dimension_tableau = selection_type_partie_console()
-        print(dimension_tableau)
-        # Attribution du nombre de bateau en fonction du style de bataille (petite ou grande)
-        nombre_bateau = selection_nombre_bateau(dimension_tableau)
-
-        # Attribution de nom de joueur
-        nom_joueur1 = nom_de_joueur(1)
-        nom_joueur2 = nom_de_joueur(2)
-
-        tableau_invisible_joueur1 = []
-        tableau_invisible_joueur2 = []
-
-        # Création des 2 joueurs et de leurs tableaux
-        joueur1 = creation_tableau_joueur(dimension_tableau, nom_joueur1)
-        joueur2 = creation_tableau_joueur(dimension_tableau, nom_joueur2)
-
-        debut_partie(joueur1,joueur2,tableau_invisible_joueur1,tableau_invisible_joueur2,nombre_bateau)
-
-    elif argument == "gui":
-        print("On démarre en mode gui")
-
-    else:
-        print("Mode de jeu inconnu.\nPrécisez une un mode de jeu en paramètre\nconsole/gui")
-
-    print("script par défaut")
